Safari-609.1.11.1/LayoutTests/crypto/workers/subtle/rsa-import-pkcs8-key-expected.txt

[Worker] Test importing a PKCS8 RSA-OAEP key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/rsa-import-pkcs8-key.js
[Worker] Importing a key...
PASS [Worker] privateKey.toString() is '[object CryptoKey]'
PASS [Worker] privateKey.type is 'private'
PASS [Worker] privateKey.extractable is true
PASS [Worker] privateKey.algorithm.name is 'RSA-OAEP'
PASS [Worker] privateKey.algorithm.modulusLength is 2048
PASS [Worker] bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS [Worker] privateKey.algorithm.hash.name is 'SHA-1'
PASS [Worker] privateKey.usages is ['decrypt', 'unwrapKey']
PASS successfullyParsed is true

TEST COMPLETE
